Exception if deleting a non-existing book

+++ src/main/java/com/polarbookshop/catalogservice/domain/BookService.java
@@ -27,6 +27,9 @@ public class BookService {
   }
 
   public void removeBookFromCatalog(String isbn) {
+    if (!bookRepository.existsByIsbn(isbn)) {
+      throw new BookNotFoundException(isbn);
+    }
     bookRepository.deleteByIsbn(isbn);
   }
